Break up and asyncify WebView initialization triggered by background threads
WebView initialization currently takes a non-trivial amount of main-thread time.
When initializing Chromium due to being on the critical path to opening a link
this is fine. However, if an application wants to pay the initialization cost
earlier to speed up a future navigation it must still pay a hefty main-threadprice,
which can cause animation jank in the app. To help mitigate this observable impact
on main thread performance, this diff introduces an alternate code path for
WebView initialization when triggered from a background thread. This code path
breaks up the higher level tasks within the intialization and enables them to
run asynchronously one after another. These tasks are scheduled by an Executor
that is synchronized with the frame rate via Choreographer frame callbacks,
which provides better guarantees that we will maintain good responsiveness
during the intialization process.